refactor: update Companies components file structure
